refactor: additional Next.js 15.1.3 upgrade changes
