+++ src/app/ClientLayout.tsx
@@ -3,7 +3,7 @@ import CssBaseline from '@mui/material/CssBaseline';
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
 import type { JSX } from 'react';
 
-import { ThemeProvider as CustomThemeProvider } from '@/styles/theme/ThemeContext';
+import { ThemeProvider as CustomThemeProvider } from '@/app/styles/theme/ThemeContext';
 
 interface ClientLayoutProps {
   children: React.ReactNode;
