+++ src/app/settings/components/PreferencesSection.tsx
@@ -1,7 +1,7 @@
 import { Box, Typography, Button, FormControl, Select, MenuItem } from '@mui/material';
 import { useState } from 'react';
 
-import { useTheme } from '@/styles/theme/ThemeContext';
+import { useTheme } from '@/app/styles/theme/ThemeContext';
 
 import type { PreferencesSectionProps } from './types';
