+++ src/app/settings/page.tsx
@@ -21,9 +21,9 @@ import { doc, getDoc, updateDoc, setDoc } from 'firebase/firestore';
 import { ref, uploadBytes, getDownloadURL } from 'firebase/storage';
 import { JSX, useCallback, useEffect, useState } from 'react';
 
+import { useTheme } from '@/app/styles/theme/ThemeContext';
 import { useAuth } from '@/lib/auth/AuthContext';
 import { db, storage } from '@/lib/firebase/config';
-import { useTheme } from '@/styles/theme/ThemeContext';
 
 import { AccountSection } from './components/AccountSection';
 import { NotificationsSection } from './components/NotificationsSection';
