+++ src/lib/providers/Providers.tsx
@@ -3,8 +3,8 @@
 import { QueryClient } from '@tanstack/react-query';
 import { type ReactNode } from 'react';
 
+import { ThemeProvider } from '@/app/styles/theme/ThemeContext';
 import { ClientAuthProvider } from '@/lib/auth/ClientAuthProvider';
-import { ThemeProvider } from '@/styles/theme/ThemeContext';
 
 import { QueryProvider } from './QueryProvider';
